+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/shared/MinimalExternalModuleDependencyExtensions.kt
@@ -5,3 +5,9 @@ import org.gradle.api.provider.Provider
 
 public fun Provider<MinimalExternalModuleDependency>.asString(): String =
     map { if (it.version != null) "${it.module}:${it.version}" else "${it.module}" }.get()
+
+public fun Provider<MinimalExternalModuleDependency>.moduleAsString(): String =
+    map { "${it.module}" }.get()
+
+public fun Provider<MinimalExternalModuleDependency>.versionAsString(): String? =
+    map { it.version.orEmpty() }.orNull?.takeIf(String::isNotBlank)