+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - `Provider<MinimalExternalModuleDependency>.asString(): String` function
+- `Provider<MinimalExternalModuleDependency>.moduleAsString(): String` function
+- `Provider<MinimalExternalModuleDependency>.versionAsString(): String?` function
 
 ### Changed